feat: show login modal event based and e2e test (#43)
As my first commit, and be being full of ideas, it introduces various
new concepts and architectural patterns to the repo that we should
discuss:

- Communication over async messages (commands & events)
- Dependency injection
- E2E testing with Playwright

All of this around the use case of showing a login modal when the user
clicks on the login button. This use case could have been done with much
less, but I wanted to show how we could do it applying the above
concepts as an example.