feat: submit register form and integration test (#50)
* feat: submit register form and integration test

For integration testing: still playing with the boundaries of the test
suite "integration". As we want it to be local and fast, but at the
same time as close to the real thing as possible.

Also adding blurr behaviour to the register form inputs. First
validation & submission is done upon "change" event, then upon "input"

Also renaming "error" to "errorMessage" as there seems to be a bug with
vue test utils where naming a variable "error" causes the test to fail.

* feat: search grid

* fix: junk

---------

Co-authored-by: Travis <travis@marketcake.com>