feat: add games module and game image (#52)
Start abstracting the games fetching to its own module, as well as
abstracting the loading of the game image to its own component.

Even if it has not been implemented yet, the game image endpoint
downloads the whole binary image, so the GamesImageLoader component
does not download it in the server nor on first print. Improving
performance and off-loading all image loading parallely to the client
after the page has been printed. A spinner appears while the image is
being downloaded.