feat: api & mock envs and first typed open api client (#53)
We set up two different environments: mock & api. That load a differnt
env file depending on what we want to run.

In the future it would be great to assert that the env types in the
index.d.ts file are correctly followed at runtime. Allowing us to
realise if we forgot to set an env varible in production.

For the Open api client: the generated client can be changed no problem,
I picked one that is super light and leverages the generated types. So
at build time it becomes super slim. But any that can be generated from
the open-api json spec, and is fully typed response & errors will work.

In the future: we should abstract the library methods with a class of
our own to become more decoupled in case we ever change to another
client.