feat: handle jurisdiction not supported (#56)
Through a middleware in the open api client, we emit an async message
to open the correct jurisdiction modal depending on the error returned
from the backend.

This way we do not have to worry about handling this error in all
backend api calls. The error ends up opening the modal automatically
as long as we use the open api client.

The alternative was to throw the error from the api client, to be
picked up at the modal layer and handled. But after some trials and
investing a bit error catching in Vue: it seemed unreliable. So we
decide to go for the async message route. The only con to this approach
is having to pass the AsyncMessagePublisher to the api client
instanciation. But it's a small price to pay for the reliability.

Also:
- moving all user-action modals to a component to de-clutter the
app.vue file.
- for now using the translation function to go from jurisdiction ("BR")
to translated country name ("Brazil").