feat: start user store and others (#58)
Starting the user pinia store with the bare minimum 
we will need. 
The store relies on a query that right now is dumb/mocked. 
In the next MR we will implement the API call to get the 
authenticated user from the client cookies.

As well as:
- starting a logger dependency. Right now the only 
implementation is a console logger. But when we go to 
production we will probably want to implement a 
logger service so we have both server and client 
logs centralised.
- unifying all places where the user sign in modal 
is opened in a method