chore: form handling improvements
Use the meta.valid property to determine if the form is valid and thus
the submit button should be disabled or not. Which allows it to be
disabled when opening the form, but without having to trigger a
validation upon form rendering.

As well as:
- using the logger dependency instead of console logs in older parts
of the code before the new logger dependency was impplemented
- moving the translation keys for validation to a more generic
"validation" key in the locales file to be re-used across different
forms.