feat: implement game categories (#62)
This commit advances the current state of showing games by categories
in the Grid components both vertical & horizontal.

Particularly:
- searching game categories by category group in home & game_page, to
then render a GridHorizontalGame component for each
- fix the category page by using the
searchGamesByCategoryPaginatingOnSlider ui use case in the
GridVerticalGames component as well
- fix the FindGameImageSrcByGameId ui use case to correctly take into
account if we are in mock or api mode so we may see the correct images
- fix some console-logs into logger-logs
- in the category back button: implemnet a composable to either
navigate back to previous page, or to the home page this is the user's
first page in our domain. This is to avoid redirecting the user outside
of our domain and potentially losing them.
- making all of the above SSR & client-side-navigation friendly
- graciously handling wrong jurisdiction errors from the backend, both
in the SSR first load as well as client-side navigation
- added the missing categories to translations. It should be noted that
this coupledness between dynamic category names and fixed translation
files is meant to break at some point. When we will add new categories
but forget to update the translation files accordingly. We should think
about how to avoid this in the future.

As well as:
- with @Tarik: fixing macOS auto fill shitty styles in the BaseInput
component