feat: add game page queries (#63)
* feat: add game page queries

Querying the game data in the games/:id page and showing it to the
User. In the next MR we will take care of loading iframe url.

The data flow approach taken is the following: favour passing down data
over props from the nuxt page downstream to the components. The reason
being. Some exceptions are made when the needed prop is needed in many
places and easy and low-cost to query (like the game image url).

What was left out of this MR:
- passing the game linked categories from game page: will implement as
soon as the backend returns them
- translation the game description & title: to define how we will
handle this
- ButtonShare functionality: details to be defined
- GameFrameVotes functionality: details to be defined
- GameFrame h3 subtitle: details to be defined

Also:
- small refactor in how we open the user action modal in favour of
reusability