feat: wallet & balance store (#64)
In a similar way we implemented the user details store, we implement
the wallet one. We implement it in a different store as, eventhough the
authentication method is the same, they are 2 separate modules
retrieved by 2 different endpoints. But for performance they are
retrieved paralelly.

A BaseCurrency component has been added to correctly display the user
current balance in all places of the frontend.

Topics to come back to in the future:
- when an non-logged in user gets to /settings/wallet, what should
happen? Right for simplicity now they get redirected to the home page.
But we might want to revisit this.