chore: fix build type errors (#66)
There is an issue where the <template> part of the vue single file
components does not recognize some auto-imported types. In our case:
the utils were not recognized.

To be more precise: the behaviour was correct, they were just
typescript error picked up by vue-tsc.

The fix that worked for me is extending the vue/runtime-core interfaces,
as described here:
https://github.com/nuxt/nuxt/issues/28869#issuecomment-2406407082

And we remove a (looks like) unused @ts-expect-error comment that
presented a build error.