feat: open the sidebar on desktop & mobile
As well as closing it automatically on page navigation.

The sidebar content is moved to the default layout as we are always
using the default one, even when using the other ones (dashboard and
carousel) as we use them as children of the default layout. This way
the sidebar open state leaves in the parent of all 3 components that
can modify it, and we can thus avoid having to emit async commands to
modify it.