chore: improve console logger
So that in the server it prints jsons that are more easily digested
by logging services. And in the frontend it still prints human readable
logs. As well as implementing a "log level" filter.

This will improve the log digestion in the server, while still keeping
a human readable log in the frontend.