chore: improve performance by using non-blocking requests (#71)
Changing the data fetching in the horizontal & vertical game grids to not happen 
in the server side rendering process, as well as be non-blocking in the client side. 
For better UX: BaseSkeleton pulsating boxes were added when first set of Games 
are loading.

As well as modifying the eslint rule for number of attributes in a single line of 
Vue Components to be more compact.