feat: add individual game provider page (#72)
Really similar to the game category page. In fact it is the exact same
design changing the game category title by the (untranslated) provider
name. For this we reuse the GridVerticalGames, as it's the same UI 
use case.

We leave the VerticalProviders component for searching providers, when
there is a use case that calls for it.

Taking the performance one step further: we are not waiting on the
server nor on the client navigation for, neither the provider name, nor
the games back. We are delaying everything so it's as snappy as
possible.