feat: add providers to search (#74)
The general component approach taken was to make the GridVerticalSearch
component agnostic to the type of items received, so it may be used to
display either games or providers depending on what is passed.

In the future we might want to refactor this and the GridVerticalGames
and GridVerticalProviders components to always use the same component
to display either Games or Providers, and abstract the backend search
capabilities. But we leave that decision for the future.

We are also using the white variant & object-contain for the provider
logos because they just look better.