chore: implement sentry (#77)
This includes replays, capturing error logs via our logger
implementation, thrown errors...

A refactor of the logger.error method was needed to pass the caused
error explicitly.

The only caveat is explained in the load-di file, where there are
problems with the sentry logger middleware on localhost. But it
technically works on production.