improv: add icon svgs to served html
So that there is no need for an HTTP call to retrieve the icons when
needed.

We choose to implement this, eventhough the bundle size will
increase, because we do not have that many icons in the first place.
So the small bundle size increase is worth avoiding the network latency
when requesting the icons, and thus the flickering the user sees while
waiting for the icons.