improve: defer register form upsert from input validation
We separate the client validation from the signup flow upsert, in order
to make the register form completion more snappy and responsive to the
user.

In the future if the signup flow upsert endpoint returns validation
errors we will have to deal with them accordingly, but for now they
are not an issue.