improv: handling error already taken in signup flow
We only handle when the backend returns an array of validation errors
per field, as this is the most common case for now for what we have
seen.

We follow the rule of reducing infra repository complexity as much as
possible in detriment of verbosity and explicit domain errors. As this
is easier to test and trace or debug.