feat: add game ratings to GameFrameVotes (#92)
At first we tried returning the ratings in the get game endpoint to
avoid a second http request. It also leads to coupling and, because
the ratings are in another backend endpoint, making the get game
request do multiple backend requests. Incrementing even more the
coupling and fragility of a core use case.

This way it's also easier to achieve a better and more friendly UX,
as refreshing the game ratings after a user has voted is at the
component level, and not at the page level.

While the game ratings component is loading, we slightly decrease the
opacity of the iccons to indicate the user that the loading is taking
place.