feat: implement locale selection (#97)
For the strategy: we rely on the nuxt i18n module to handle 
the locale setting, translations, and locale state. While we 
wrap the slightly complex logic of determining the user 
locale in the way we want to, and locally saving the user 
selection in localstorage. 

We use this same logic to pass to the backend on user 
signup submission their preferred locale.

We change the current locale from "en" to "en-US" to be 
more explicit and follow better locale practises. We take 
care of setting the necessary fallbacks and rules so that 
other "en-XX" users will fall into "en-US".

We also change how the Flag component works to 
leverage the iconify custom collections. This way we can
have hundreds of flags and the icon module takes care 
of optimising and including them in the build as icons.

In a next commit we will change the flag filenames to 
lowercase to remove the iconify warnings.