feat: implement path-based locales (#99)
The full explanation and description is in the shared google doc in the
related PR #99.

The summary of the intended UX behaviour is the following:
A new user that has not manually selected a locale in the footer will
be redirected to their browser's locale (if supported) when landing on
the root, locale-less, path. If landing on a locale-prefixed path: they
will stay in that locale.

A user that manually selects a locale in the footer will always be
redirected to that locale.

We lowercase the locales (en-US -> en-us) for better readability in the
url.

Other miscellaneous changes:
- a BaseLink component is introduced to avoid future issues with changing
all internal links, like it happened here with the need to change
NuxtLink for NuxtLinkLocale.
- in order to better implement the footer language selector, we
need to change the BaseSelect component to allow the usage of v-model
with the selected option.
- We copy-paste the en-US.json to pt-BR.json to have a second locale to
test the locale redirect, and we copy all the translations to the new
one to not be flooded with warns when switching to the pt-BR one.