feat: open websocket connection with backend
This commit is only for opening it from a client Nuxt plugin. When the
format of the messages to receive from backend is more clear, we will
decide on how to handle messages on a use-case-per-use-case basis.

For the re-connection policy, knowing that a lease token is needed to
establish a connection, and that they are valid for only 60 seconds,
when a connection is lost because of an error we create a new
connection with a new token. For simplicity.

In the future, depending on how we will handle the websocket messages,
we might want to move the websocket connection to a vuex store so it
can be accessed from multiple components.