feat: dynamic deposit max min amounts
As well as making the deposit modal be dependent on the user payment
method id.

The approach taken was to query the backend data from the user payment
method in the ModalDeposit once the client is loaded, in a non blocking
way. Then it gets passed down as props to the FormDeposit.

This will allow us in a future when we have multiple payment methods or
multiple currencies to implement some choosing logic in the
ModalDeposit without conditioning the Form.