feat: update user settings consents (#111)
Handling the consents is a bit more complex than initially planned,
as the backend update endpoint differenciates between: undefined, null,
and a boolean value. But the frontend use case only cares about the
boolean.

We differentiate the UserSettings naming the Backend uses, from the 
"Simplified" user settings consent we decide to show in the frontend. 
Which groups some of the consents together.

We introduce a new BaseToggle component, the design is not final but
good enough to start with. For the component API we decide to keep it
simple, we will be able to add more complex props and config when
needed.

Regarding the loading of the user settings store: I decided to wait
until the user visits this page to load this data (opposed to lazily
waiting for the frontend to load it on any page the user lands in),
because we are starting to accumulate quite a bit of requests to the
backend when the client finishes loading. And I believe this one will
be needed in 0,0001% of cases. So IMO it's better to make those little
users wait for a skeleton than drown the backend with requests that
will be 99,999% not needed in most cases