feat: receive websocket payment message
For now we implement a simple architecture for the WS messages:

WS message is received and handled in the websocket client plugin,
which then emits an async message through the async message publisher.

In the case of the "payment_status_update" message: it ends up causing
the wallet module to refresh, and thus show the possibly changed
balance.