feat: deposit confirm modal functionality (#113)
Right now we are passing all props to the DepositConfirm modal, so they
can be passed to the Form DepositConfirm. Seeing the number of props,
this might not be the best approach, an alternative would be quering
from the backedn all flow details from a GET flow endpoint.
But until we have that endpoint in the backend, we will keep this
approach.

We refactor a bit the BaseCopy component to be more flexible, we would
like to use it in all use cases where we need to copy a text by
clicking on something, seeing some kind of visual feedback after
clicking.

As well as closing the modal when payment with same flow_id is received
through WS message.