fix: currency inputs in deposit and withdrawal
By implementing more complex masking, we can safeguard the user from
entering wrong characters.

To achieve this now we can input mask options as well as mask pattern
in the Input and InputGroup components.

Because maska returns and receives strings, a simple way to deal with
it is to convert the value to a number when the user inputs it. Loosing
the 2-way v-model binding, but in these use cases we do not need it.

This fixes #112