fix: change prefetch to preload
Because browser does not care about the link prefetch and was not
downloading the images.

The only con about using preload over prefetch is making the user
download images that may not be needed, but we can always change this
in the future in case we see it becomes a problem.