improv: make wallet request async in client
To improve first page load time. Now the only remaining blocking
request is the user authentication request, which we decide to maintain
because it does change the navbar UI significantly.

We also implement a small skeleton in the balance navbar to reduce UI
shifting when the wallet balance is loaded.