feat: changing user setting time zone and language
From the settings/account page.

We have taken a "fire and pray" approach to the change of the user
settings to increase the feeling of snapiness on non-critical
operations. This means that the modal closes right when the user clicks
on save, supposing that the request will be successful. In the
unlikily case of an error: it will be sent to Sentry, but the user will
not see it went wrong.

We also change the BaseForm to pass on the loading boolean as a slot
prop. Which will clean out a bit the necessary code to use the base
form.