feat: v1 combobox in update account TimeZone
In order to not loose too much time pondering on a thousand details, I
followed these criteria:
- make it look like the BaseSelect
- as simple as possible to implement
- do not strive away from Headless UI practises to keep it stable and
accessible

The biggest thing to talk about IMO is the decision to, when the
options are open and visible, make their position absolute or block.
Usually I would have gone with absolute like in the BaseSelect, but
this use case places the combobox in a *small* modal. Making a big
list of options awkward to scroll through. So I set it as block for now,
but can be revisited, even set as a prop depending on where we use
the component.