fix: attempt to fix the user telephone codes
For some reason: the country codes used in the UserTelephone class
are being correctly imported from the client, but not in the server.
Resulting in the user authentication store failing in SSR but not in
the client.

This commit attemps to fix it by using named imports instead of default