fix: nav user icon centered
I decided to remove the css height as we already set a size with the
icon component, and it was only hindering the centering of the icon.

This made the navbar a bit shorter, as it was being affected by the
height of the icon, if we want to make it a bit taller again: we
should specify it at the navbar level, not the icon.

This fixes #119