feat: kyc sumsub flow and titanic refactor of modal
In order to see it: head to the account > verification tab. If the
current user has not completed the KYC process, a button will appear
to open the modal that will load teh sumsub SDK frame.

The refactor was to simplify the modal state and how the types are
shared, to reduce layers and duplicity.

As well as introducing in the common dependencies a "requestContext"
to easily inject request-bound data into the DI easily.