fix: modal images preloading
To fix it we had to preload the exact same image src Nuxt uses to
display the image in the modal. This way the image is already cached
when the modal is opened and the image is displayed instantly.

Some warns appear in the console as "preload" should only be used
with resources directly needed for the first page load, but we accept
them for now as using "prefetch" does not download the images at all.

This fixes #121