improv: add spinner while KYC iframe loads
In production we do see a good 1-3 seconds before the SDK loads the
first loading animation. During those seconds the user sees a
completely empty modal. If impatient, they could think that something
broke and close the modal.

To prevent this, we place a loading spinner until the Sumsub SDK loads.