feat: get games from category list edp on sliders
Now that the list categories edp can return some of the games of each
category, we take advantage of it in the homepage to show the user the
home categories alongside a first selection of games, all in one
request. This should improve the time needed to show the games to the
user.

This did present the question on how to match the initial games
provided by the categories endpoint with the pagination handled by the
slider. The approach taken was to pass the initial games as props, show
them as soon as the component renders, and then fetch the games in
the background. When they are fetched: switch the initial games for the
queried ones. This way the user can see the initial games as soon as
possible, and then the rest of the games will be shown as they are
queried.

In order to make this and the vue component update the games
accordingly, we had to change the BaseSlider generic to need a "key"
string attribute that the component uses to keep track of each item.
This is necessary now that the items array can change, and allows for
more flexible usage and modification of the slider and its items.

This solves #123