improv: WS reconnection logic
Now the WS attempts to reconnect up to 8 times with an exponential
delay. The exponential delay strategy was chosen to avoid a massive
linear spam of reconnections in case the server disconnects a big
number of users (a server restart, for example).

Also, when reconnecting it attempts to re-open the logged-in user
WS channel