feat: Notifications module & save closed PromoBar
Until we decide what to do with the problem of translation in banner
notifications coming from the Backend, we generate them from the code
(NotificationRepositoryLocalStorage) using the local translation json
files. The repo uses localstorage to know if the user has marked the
banner as read or not. And adding or modified banners can be done from
the repo and the translation files.

Next step will be implementing a NotificationRepositoryBackend that
will fetch the notifications from the backend and show them as toasts.
The frontend will leverage both repositories until we have a final
solution for the banner i18n issue.