fix: attempt to fix sentry server config
As of 2025-01-01, the Sentry server-side tracing is currently still in
beta, all of a sudden we started having runtime errors
"ReferenceError: require is not defined in ES module scope, you can use
import instead" And the only way to fix this without downgrading the
package version too much was to deactivate Sentry server-side.

(I tried deactivating traces only to keep the error reporting, but the
same reference error stayed)

My recommended approach would be waiting until Sentry gets their shit
togheter and go out of beta, or at least wait until right before our
launch and see the state they are in.

The commented code left here was working error-capturing-wise as well
as sending traces to Sentry before the error started showing up.
But the OTEL config to send the traces to our Grafana was not still
working. By the time we un-comment it, config will have probably
changed.