feat: notification toast on WS message
For now we leave the styles really similar to the BaseAlert component.

We differenciate between Banners and Toasts as, for now, the plan is
to generate banners statically from the frontend to cover for
flexibility and translations.

Left to do: when user opens site pull the unread notifications from the
backend and show them as Toasts. Also: mark as read the notifications
received from the WS messages so the user does not see them again after
reloading the page.