feat: mark notification toasts as read when shown
More specifically: when cleared from the window, in case a user leaves
the page before the toast is cleared.

Note: I need to refactor the websocket-messages interface to not have an
additional abstraction layer between the server WS messages and our
internal async messages. There is just no need.

In an effort to reduce boilerplate, we use the generic apiBaseUrl env
variable as the notifications backend url. If we ever need to switch
off only the notifications backend repository: we will implement a
apiBaseUrl just for it.