feat: show unread notifications on page load
In the future when the backend allows it, we will only query the
last X hours of unread notifications. Now we are querying the last 7.

The abstraction layer between all WebsocketServerNotifications and the
NotificationToast we show the user is in the "NotificationToastBuilder"
class. If we want to update, add, or remove notifications we can do it
from there.

As well as simplifying the typing of backend notifications. But it
could be simplified a step further.