feat: implement Live Chat widget
the external scripts needed are loaded only once during navigation,
and are triggered lazely once the dom is ready. This is done by using
dynamic nuxt components.

In order to prevent the horrible hovering widget in the corner from
always being there: we hide it when minimized.
When the user clicks on any button that opens the widget: it gets
maximized so the user can chat. And when the user closes the chat:
the widget completely disappears.

Left to do:
- Look into the widget language and how to customise it.
Right now it automatically selects the language based on IP (?)
- Pass to the widget the currently logged-in user information/data so
the agent knows which user they are talking to

Progress towards #125