feat: vertical hover animation on HorizontalGames
As requested by Chad, we implement a slight vertical animation.

In order to do so: we had to create a GameImageLink component that
introduces a small vertical margin so the image has the space to be
translated on hover.

The component allows to select which animation to apply, for now we
kept the previous zoom-in animation for all vertical greads (category
page and search results)

The only thing I'm not happy about is having to reduce the gap-4 to
gap-2 on the GridHeader component when used in the GridHorizontalGames.
This is to balance out the introduced top margin necessary for the
vertical translate. In a perfect world this should be achieved from
inside the GameImageLink component to reduce coupling, but I could not
find a better way for now.