fix: deposit modal buttons going over-width on xs screens
As pointed out by Chad, in really small screens (under 400px), the
deposit form buttons were stacking and going off-screen.

I added a new viewport size of 400px "xs" to the tailwind config, so
that we may specify styles for even smaller screens, while setting the
default style for "xs and above" as we have been doing. The idea would
be to do this really sparsingly, only like these case where the UI
breaks down because the screen is too small.