fix: game search endpoint spam
While testing some other feature, I saw that using the `useThrottleFn`
the search endpoint was being spammed at every key press, no matter
how fast it was done. So I changed it to the other approach we tried in
the past with `debouncedWatch` and now the endpoint is only being
called once at the end of the delay (150ms) with the latest query.

We keep this for now and will be testing it with the team