feat: add responsible gaming and T&C pages translated
Following last commit's way of handling translations for long pages

The only thing I am not happy about is how, in the terms.vue and
responsible-gaming.vue pages, the number of sections is hardcoded.
It should somehow auto-detect how many sections there are in the
translation files and render them accordingly.