feat: GameAction type refactor & MyBets tab in homepage
As requested by Chad, we implement a "my bets" tab in the "winning now"
section of the homepage. In order to do this, we change the data table
components into DataTableGameActionsCasino and
DataTableGameActionsMyBets. So we can have different tables but with
shared translations and placed together. Both using the same UI handler.

We take this change to refactor the GameAction type to use the return
type from the backend in order to simplify layers in exchange for a bit
more coupling.