improv: reduce gap spacing in mobile
So that, as requested by Chad, most users can see the second row of
games in the home page. Spacing in desktop has been kept pretty
similarly.

I also refactored it so more components use the --giro-section-gap
CSS variables and all sections are more consistent.

If we decide it's too tight in mobile, we only need to incrase the
--giro-section-gap-sm variable to the desired value (before: 2rem).