feat: allow custom CSS variables for sites
In the spirit of keeping the "separate assets folder for the different
sites", we add a `variables.css` file for each site. Which loads some
css variables that are used by the tailwindcss classes.

This way, when adding a new site, we just need to add the
`variables.css` alongside the rest of assets.