feat: send utm params on signup
Now we save the utm parameters in the query parameters when a user lands
on the frontend. Then, eventually when they sign up, we send them to
the backend alongside the rest of signup data.

We also remove the utms from url to avoid clutter and the user
copy-pasting the url to share it and taking the params alongside.

Some of these changes are bound to change, for example if Chad decides
to change the returned params to the last ones the user landed on the
site with. Or when Daniel changes how we send the params to the backend
upon signup.