feat: partially implement GTM
As explained over slack:

Changes done in GTM itself:
- created all GTM triggers as per what you requested on whatsapp
- added the data layer variables we will be using in the tags
- @nude for now I've only added a couple of tags that allow me to check
everything is working as expected. I believe you & will be creating
the tags with the pixels & everything you want right?
Dunno which ones to create without the pixel ids, but I can do it if
you tell me

Changes in commit: I've implemented the following triggers in the code
correctly and checked they fire using the GTM debugger. But I believe
I need Daniel's help to trigger the following ones:
BET_FINALIZED, DEPOSIT_FINALIZED , WITHDRAWAL_FINALIZED. As well as
knowing upon DEPOSIT_INITIATED if it's the first ever deposit of the
user or not. Perhaps I already have the necessary info in the frontend,
but I would like to check it with you.