fix: use identifiers instead of numeric ids
This commit has a ton of changes as most of the code was dependent on
on numeric ids. From types to variable names.

In a future refactor we will try passing the whole game type down the
components tree instead of passing game properties separately.
This will make the frontend data flow easier to manage.

We also take the change to refactor some code to be more tied to the
backend types.