feat: record clicks in tracking system
Upon user landing on the site we call the tracking system to record the
click associating it with user's fingerprint. As well as linking the
tracking long lived session with the user_id upon login & signup.

All of this is done in the clicks-tracking-init.client.ts file, to keep
all tracking-related code in one place.

If we want to interact with the tracking system locally, we need to
change the .env variable "NUXT_PUBLIC_TRACKING_API_BASE_URL" to point
to the tracking server address.