improv: refactor game page so that concurrent requests
The new LoadItemsWrapper component allows to easily render multiple
items fetched from an API without having to call the fetch in the
parent's setup function. This allows to call the fetch in parallel
with other child components.

As well as moving all GameFrame... components to a new folder for
better organization