fix: duplicate wins in homepage
Now using a map with the win ID to ensure no duplicates are displayed.
We are still querying the endpoint when user re-lands in the homepage,
to ensure we are displaying the most recent wins given by the API. This
is because we are abandoning the WS channel when the user leaves the
homepage, thus loosing the live updates while they are away from the
homepage.