fix: slighly the new Game NuxtPicture implementation
By applying css background image trick to show a placeholder in case
the image is not found. But I am not too happy with this solution
still, I will let sit it for a bit and if not we can go back to
creating the <picture> and <source> elements manually.