feat: add alt text to game images not found
It's ugly, but right now there are 50% of missing images, and the
testing users have NO way of knowing which games appear when searching.

The alt text at least will tell them the title of the game.

This could be re-implemented in a better way to show the game title
beneath the brand icon in a better way, but for now this is a quick fix.