!chore: update to nuxt 4
Minor adaptions to make it work with Nuxt 4.

Biggest adaption was dropping the Google Tag Manager plugin
"@zadigetvoltaire/nuxt-gtm" because it is not compatible with Nuxt 4
and it seems to be abandoned by their authors.

Instead we install the "@gtm-support/vue-gtm" package, which is the one
used underneath the dropped one, and import it directly from our
client plugin that uses it.