hotfix: prevent register form submission race condition
Before this hotfix, with poor a network and a fast user, the user could
attempt to submit the form before the validation and updating of the
last field was completed, leading to the submission of the form
with missing data.

By waiting for the loading of each field to complete before allowing
the form submission to proceed, we can avoid potential race conditions
and ensure that all field validations are properly handled.