hotfix: register form race condition 2nd take
After the last hotfix the out-of-sync issue has happened once more.

In an effort to be exhaustive, when the user presses the submit button,
we update the submit flow one last time with the current values of the
form. Ensuring that the values the form submits with are the ones the
user sees on screen (that have already been validated).