feat: rework of Websockets client
Summary of the rework:
- simplified working with user-related channels ("user" & "tracker")
so that components and other frontend parts can subscribe to messages
without having to know about access tokens, order of joining channels,
which messages belong to which channel, etc...
- separating channel managers between user-related and open channels.
Both easily link user-related state hooks to the its channel manager,
and to easily manage joining and leaving the winning_now channel so
that the backend does not have to continously send winning_now messages
when the user does not need them.

Pending to do: verify how does the frontend respond to long-lasting
windows and WS reconnection.