!!update sentry package
Historically this has brought some issues and lack of reporting server
side. After commit I will check we are still receiving traces & errors
in production

Added pinia store integration to see the current state of the stores
when an error is reported.

As well as attempting to add the release version to the
server-side Sentry configuration.