feat: refactor SignupFlow upsert with improvements
- before we were sending all previous and changed data upon input
change, now we only send the input that has changed. Allowing for
multiple calls to be made at the same time without risking ones
overriding the others nor race conditions.
- before we made sure the signup flow existed before patching it,
now we patch it directly to improve snapiness.
- before we waited for the first input change to create the signup flow,
now we create it in a client plugin when the app initializes if the
user is not authenticated, so that the signup flow is already created
when the user starts typing in the form.
- before we re-validated the fields right before patching the signup
flow, now we pass them as they are to reduce complexity. It should
not be an issue as we only call the patch method after successful
validation.