improv: not perform any validation on user client-side
Before we validated if the user's phone was a valid phone. But this is
not a good practise since we prefer to leave these validations to the
server.

Now we only return the bare backend user type with primitive values.
And when we need to interact with this phone (and its validity), we do
by setting some friendly fallbacks in case it is not.